fix(mcp): name the stale-grant case, and share the tool-result shape
Two review notes, both real.

A Hub deleted after consent leaves its uuid on the token, so the granted set is
non-empty while the query returns nothing. That fell through to 'several Hubs
are available and none is open — call pluggedin_open_hub', which is false on
both counts: none are available, and the tool it recommends fails the same way.
A wrong message that recommends a dead end is worse than no message. It now says
the granted Hubs no longer exist and to re-authorize, which is the only thing
that helps.

ToolResult and its text/failure helpers were copied between the hubs and library
handlers, with four more groups to come. Moved to lib/mcp/connector/tool-result.
Duplicated response shaping drifts quietly — one file starts pretty-printing,
another stops setting isError, and a model sees two conventions from one server.

Mutation: removing the stale-grant branch fails the new test.
